Make most fields of StrobemerIndex private
